Add working implementation for condition-variable-wait on ECL
Until now if timeout argument was provided ECL signalled a simple-error that
timedwait is not supported for CV. We wrap ordinary mp:condition-variable-wait
in with-timeout to make it work.

This fix is required to make generic implementation semaphores work on ECL prior
to 16.2.0. Thanks to that all tests pass (while previously tests involving timed
wait on CV failed).

We make default with-timeout implementation available for
implementation-specific code to use.