Added null-impl for apiv2 to fix loading on CLISP with no threads support.
Probably this will fix loading on other systems too, when there is not
threading support.

In old BT version implementation files were included condiitionally
if there is implementation name and :thread-support in features. But
when apiv2 was introduced, it dropped check for thread support.

This pull returns this check back.