Provide a proper implementation of the V2 API for Genera.
Move QUEUE from atomics.lisp to impl-condition-variables-semaphores.lisp, which is
the only file which uses QUEUE, as the Genera implementation does not define %WITH-LOCK.
(impl-condition-variables-semaphores.lisp is only used by CCL.)

On Genera, fix the calls to PROCESS:PROCESS-KILL in DESTROY-THREAD --
the WITHOUT-ABORTS argument is an optional not a keyword argument.

Fix a race condition in locking and unlocking --
Assume two processes, A and B, and a lock. Process A locks the lock. Process B then tries
to lock the lock and blocks until A unlocks it. As currently written, when process A
unlocks the lock, Genera might switch to process B before process A can reset the lock's
internal BT state to indicate it no longer has the lock. Process B then updates the
internal state to indicate it has the lock which Genera has given it. Process A then
runs and reset the lock's internal state. When process B tries to unlock the lock,
it will crash because it passes an invalid value to Genera's process:unlock.

The fix is to make all lock/unlock operations in BT atomic w.r.t. locking/unlocking
the Genera lock and updating the internal state.